requirements.txt
playwright
fastapi
pillow
pytesseract
uvicorn
python-multipart